fix: cast psgc_id to INTEGER in batch UPDATE query
PostgreSQL VALUES clause treats all values as TEXT by default.
Added ::integer cast to psgc_id to match column type.

Error: column "psgc_id" is of type integer but expression is of type text

+++ src/routes/clients.ts
@@ -995,7 +995,7 @@ clients.post('/psgc/assign', authMiddleware, requirePermission('clients', 'updat
       await pool.query(`
         UPDATE clients AS c
         SET
-          psgc_id = v.psgc_id,
+          psgc_id = v.psgc_id::integer,
           region = COALESCE(v.region, c.region),
           province = COALESCE(v.province, c.province),
           municipality = COALESCE(v.municipality, c.municipality),
